created implementations for some services
