+++ src/test/java/com/team4/testingsystem/security/JwtRequestFilterTest.java
@@ -1,6 +1,6 @@
 package com.team4.testingsystem.security;
 
-import com.team4.testingsystem.services.CustomUserDetailsService;
+import com.team4.testingsystem.services.impl.CustomUserDetailsService;
 import com.team4.testingsystem.utils.EntityCreatorUtil;
 import com.team4.testingsystem.utils.jwt.JwtTokenUtil;
 import org.junit.jupiter.api.Assertions;
