+++ src/test/java/com/team4/testingsystem/services/impl/CustomUserDetailsServiceTest.java
@@ -1,4 +1,4 @@
-package com.team4.testingsystem.services;
+package com.team4.testingsystem.services.impl;
 
 import com.team4.testingsystem.entities.User;
 import com.team4.testingsystem.entities.UserRole;
